basic routing, still kinda buggy but fine
