+++ app/src/config.ts
@@ -1,0 +1,1 @@
+export const base_api: string = import.meta.env.API_URL || 'http://localhost:3000/';
