+++ app/src/routes/models/+page.svelte
@@ -1,0 +1,7 @@
+<Navbar/>
+<Models/>
+
+<script lang="ts">
+    import Navbar from '../Navbar.svelte';
+    import Models from './Models.svelte';
+</script>
